+++ src/toll/run.py
@@ -13,21 +13,37 @@ class Runner:
     def __call__(self):
         for command in self.commands:
             for package in self.packages:
-                if not self.run(command, package):
+                if not self.run(
+                        command.command, command.precondition, package):
                     self.render_failure()
                     return False
         self.render_success()
         return True
 
-    def run(self, cmd, package):
-        print('Running', cmd, 'on', package)
+    def run(self, cmd, precondition, package):
         cwd = os.getcwd()
         os.chdir(package)
         try:
-            process = subprocess.Popen(cmd.split())
-            return process.wait() == 0
+            if self._precodition_is_met(precondition):
+                print('Running', cmd, 'on', package)
+                return self._run_cmd(cmd)
+            else:
+                print('Not running', cmd, 'on', package)
+                print('Precondition', precondition, 'on', package, 'not met.')
+                return True
         finally:
             os.chdir(cwd)
+
+    def _precodition_is_met(self, precondition):
+        if not precondition:
+            return True
+        if self._run_cmd(precondition):
+            return True
+        return False
+
+    def _run_cmd(self, cmd):
+        process = subprocess.Popen(cmd.split())
+        return process.wait() == 0
 
     def render_failure(self):
         print('FAILURE :-(')