+++ src/toll/test_main.py
@@ -1,6 +1,6 @@
 from .compat import mock
 from .main import main
-from .testing import get_packages, test_command
+from .testing import get_packages, raw_test_command
 import os.path
 import pytest
 import re
@@ -18,11 +18,14 @@ def toll_ini(tmpdir):
             package_paths.append(dest)
         ini = tmpdir.join('toll-test.ini')
         ini.write('''\
-[commands]
-test = {test_command}
+[test]
+command = {test_command}
+[test2]
+precondition = test -e fine.py
+command = {test_command}
 [packages]
 {packages}
-'''.format(test_command=test_command, packages='\n'.join(package_paths)))
+'''.format(test_command=raw_test_command, packages='\n'.join(package_paths)))
         return str(ini)
     return config_file
 
@@ -46,3 +49,15 @@ def test_main__main__2():
             # toll.ini file is not valid but that's okay for the test
             pass
     assert 'toll.ini' == parsed_file.call_args[0][0].name
+
+
+def test_main__main__3(toll_ini, capsys):
+    """It omits commands whose precondition is not met."""
+    config_path = toll_ini('bad', 'fine')
+    main(['-c', config_path, 'test2'])
+    out, err = capsys.readouterr()
+    assert re.search('^Not running .* setup.py -q test on .*/bad$', out, re.M)
+    assert re.search(
+        '^Precondition test -e fine.py on .*/bad not met.$', out, re.M)
+    assert re.search('^Running .* setup.py -q test on .*/fine$', out, re.M)
+    assert out.strip().endswith('SUCCESS :-)')